Address reviews - move adjustPositionForBedrock to CollisionManager
